% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Summarize a simloglm object}
\usage{
get_summary(input_obj, which_qoi = "median", alpha = 0.05)
}
\arguments{
\item{input_obj}{An object of class simloglm (the output of calling simloglm).}

\item{which_qoi}{Are you interested in the conditional geometric mean (which
is also the conditional median, \code{"median"}) or the conditional arithmetic
mean (\code{"mean"})? Default is \code{"median"}.}

\item{alpha}{The significance level for the resulting confidence interval.
Default is 0.05.}
}
\value{
A list of class "summary_simloglm" with the point estimate for every
scenario and the corresponding simulation quantiles.
}
\description{
Summarize a simloglm object
}
\examples{
df <- cars
regression <- lm(log(dist) ~ speed, data = df)
# Specifying no scenario simulates at the mean of speed.
simulation_results <- simloglm(regression, verbose = FALSE)
simulation_summary <- get_summary(simulation_results)
simulation_summary
}
