% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hollibaugh_rothenberg_2018}
\alias{hollibaugh_rothenberg_2018}
\title{The dataset for the Hollibaugh/Rothenberg 2018 replication}
\format{
A data frame with 9382 rows and 70 variables.
}
\source{
\doi{10.7910/DVN/B9SCDJ}
}
\usage{
hollibaugh_rothenberg_2018
}
\description{
For more information see the documentation in the dataverse of Hollibaugh and Rothenberg
}
\keyword{datasets}
