% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_difference.R
\name{get_first_difference}
\alias{get_first_difference}
\title{Get the first difference from a simloglm object}
\usage{
get_first_difference(input_obj, which_mean = "geometric_mean", alpha = 0.05)
}
\arguments{
\item{input_obj}{An object of class simloglm (the output of calling simloglm). With two scenarios.}

\item{which_mean}{Are you interested in the conditional arithmetic mean ("arithmetic_mean") or conditional geometric mean ("geometric_mean"; same as median)? Default is "geometric_mean".}

\item{alpha}{The significance level for the resulting confidence interval. Default is 0.05.}
}
\description{
Get the first difference from a simloglm object
}
\examples{
df <- cars
regression <- lm(log(dist)~speed, data = df)
# Explicitily specifying two scenarios.
simulation_results <- simloglm(regression, scenario = list(speed = c(5, 20)))
first_difference_summary <- get_first_difference(simulation_results)
first_difference_summary
}
