% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratio.R
\name{get_ratio}
\alias{get_ratio}
\title{Get the ratio between two scenarios from a simloglm object}
\usage{
get_ratio(
  input_obj,
  which_qoi = "median",
  which_scenarios = c(1, 2),
  alpha = 0.05
)
}
\arguments{
\item{input_obj}{An object of class simloglm (the output of calling simloglm). With at least two scenarios.}

\item{which_qoi}{Are you interested in the median or mean ? Default is "median".}

\item{which_scenarios}{If you pass a simloglm object with more than two scenarios please specify which scenarios should be used. Default is c(1, 2).}

\item{alpha}{The significance level for the resulting confidence interval. Default is 0.05.}
}
\description{
Get the ratio between two scenarios from a simloglm object
}
\examples{
df <- cars
regression <- lm(log(dist)~speed, data = df)
# Explicitily specifying two scenarios.
simulation_results <- simloglm(regression, scenario = list(speed = c(5, 20)))
ratio_summary <- get_ratio(simulation_results)
ratio_summary
}
