% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Summarize a simloglm object}
\usage{
get_summary(input_obj, which_mean = "geometric_mean", alpha = 0.05)
}
\arguments{
\item{input_obj}{An object of class simloglm (the output of calling simloglm).}

\item{which_mean}{Are you interested in the conditional arithmetic mean ("arithmetic_mean") or conditional geometric mean ("geometric_mean"; same as median)? Default is "geometric_mean".}

\item{alpha}{The significance level for the resulting confidence interval. Default is 0.05.}
}
\description{
Summarize a simloglm object
}
\examples{
df <- cars
regression <- lm(log(dist)~speed, data = df)
# Specifiying no scenario to simulate at the mean of speed.
simulation_results <- simloglm(regression)
simulation_summary <- get_summary(simulation_results)
simulation_summary
}
