% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.simloglm}
\alias{plot.simloglm}
\title{Plot function for a simloglm object}
\usage{
\method{plot}{simloglm}(input_obj, which_mean = "geometric_mean", alpha = 0.05, twosided = TRUE)
}
\arguments{
\item{input_obj}{Either a list of class "lm" (the output from a call to lm) or a user provided list with the following entries beta_hat (the estimated regression coefficients),
varcov_hat (the estimated variance covariance matrix), sigma_hat (the estimated residual standard error), n (the number of observations) and k (the number of regression coefficients).
The list can be provided for more flexibility. Most users will call simulate directly on the output from a call to lm.}

\item{which_mean}{Which mean should be plotted? Geometric mean ("geometric_mean") or arithmetic mean ("arithmetic_mean"), defaults to "geometric_mean".}

\item{alpha}{Set the significance level. Default is 0.05.}

\item{twosided}{Do you want two sided confidence intervals? Defaults to TRUE.}
}
\value{
A plot of the simloglm object.
}
\description{
Plot function for a simloglm object
}
\examples{
df <- cars
regression <- lm(log(dist)~speed, data = df)
# Specifiying no scenario to simulate at the mean of speed.
plot(simloglm(regression))
# Explicitily specifying a scenario.
plot(simloglm(regression, scenario = list(speed = c(5, 10, 20))))
}
