% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shepherd_you_2020}
\alias{shepherd_you_2020}
\title{The dataset for the Shepherd/You 2020 replication}
\format{
A data frame with 3074 rows and 90 variables.
}
\source{
\url{https://doi.org/10.7910/DVN/IN6SFG}
}
\usage{
shepherd_you_2020
}
\description{
For more information see the documentation in the dataverse of Shepherd and You.
}
\keyword{datasets}
